some improvement code level
