+++ backend/src/__test__/controllers/students/removeStudent.test.ts
@@ -1,48 +1,46 @@
 import { type Request, type Response } from 'express';
-import { Student } from '../../models/student';
-import AppDataSource from '../../dataSource';
-import { updateStudent } from '../../controllers/student.controller';
+import { Student } from '../../../models/student';
+import AppDataSource from '../../../dataSource';
+import { removeStudent } from '../../../services/student.services';
 
-jest.mock('../../dataSource', () => ({
+jest.mock('../../../dataSource', () => ({
   getRepository: jest.fn().mockReturnValue({
     findOne: jest.fn(),
-    merge: jest.fn(),
-    save: jest.fn()
+    remove: jest.fn()
   })
 }));
 
-describe('updateStudent', () => {
+describe('removeStudent', () => {
   let req: Partial<Request>;
   let res: Partial<Response>;
 
   beforeEach(() => {
     req = {
-      params: { id: '1' },
-      body: { name: 'John Doe' }
+      params: { id: '1' }
     };
     res = {
       status: jest.fn().mockReturnThis(),
-      json: jest.fn()
+      json: jest.fn(),
+      end: jest.fn()
     };
   });
 
-  it('should update a student', async () => {
+  it('should remove a student', async () => {
     const mockStudent = { id: 1, name: 'John Doe' };
     const studentRepo = AppDataSource.getRepository(Student);
     (studentRepo.findOne as jest.Mock).mockResolvedValue(mockStudent);
-    (studentRepo.save as jest.Mock).mockResolvedValue(mockStudent);
 
-    await updateStudent(req as Request, res as Response);
+    await removeStudent(req as Request, res as Response);
 
-    expect(res.status).toHaveBeenCalledWith(201);
-    expect(res.json).toHaveBeenCalledWith(mockStudent);
+    expect(res.status).toHaveBeenCalledWith(204);
+    expect(res.end).toHaveBeenCalled();
   });
 
   it('should handle student not found', async () => {
     const studentRepo = AppDataSource.getRepository(Student);
     (studentRepo.findOne as jest.Mock).mockResolvedValue(null);
 
-    await updateStudent(req as Request, res as Response);
+    await removeStudent(req as Request, res as Response);
 
     expect(res.status).toHaveBeenCalledWith(404);
     expect(res.json).toHaveBeenCalledWith({ error: 'student not found' });
@@ -54,7 +52,7 @@ describe('updateStudent', () => {
       new Error('Test error')
     );
 
-    await updateStudent(req as Request, res as Response);
+    await removeStudent(req as Request, res as Response);
 
     expect(res.status).toHaveBeenCalledWith(500);
     expect(res.json).toHaveBeenCalledWith({ error: 'internal server error' });
